+++ src/main/java/hu/kazocsaba/math/matrix/Vector.java
@@ -19,12 +19,26 @@ public interface Vector extends Matrix {
 	 */
 	public double getCoord(int index);
 	/**
+	 * Retrieves the specified coordinate of the vector without checking the argument. If the index is invalid,
+	 * the behaviour of this method is unspecified. For example, it may return some value or throw an exception.
+	 * @param index the coordinate to return
+	 * @return the same as {@code get(index,0)}
+	 */
+	public double getCoordQuick(int index);
+	/**
 	 * Updates the specified coordinate of the vector.
 	 * @param index the coordinate to set
 	 * @param value the new value
 	 * @throws IndexOutOfBoundsException if the index is out of range
 	 */
 	public void setCoord(int index, double value);
+	/**
+	 * Updates the specified coordinate of the vector without checking the index. If the index is invalid,
+	 * the behaviour of this method is unspecified. For example, it may set some element or throw an exception.
+	 * @param index the coordinate to set
+	 * @param value the new value
+	 */
+	public void setCoordQuick(int index, double value);
 	
 	/**
 	 * Returns the dot product of this vector and the argument